Change back to int

+++ src/bin_heap.h
@@ -4,58 +4,58 @@
 using namespace std;
 
 // Prototype of a utility function to swap two integers
-void swap(long long *x, long long *y);
+void swap(int *x, int *y);
 
 // A class for Min Heap
 class MinHeap
 {
-	long long *harr; // pointer to array of elements in heap
-	long long capacity; // maximum possible size of min heap
-	long long heap_size; // Current number of elements in min heap
+	int *harr; // pointer to array of elements in heap
+	int capacity; // maximum possible size of min heap
+	int heap_size; // Current number of elements in min heap
 public:
 	// Constructor
-	MinHeap(long long capacity);
+	MinHeap(int capacity);
 
 	// to heapify a subtree with the root at given index
-	void MinHeapify(long long i);
+	void MinHeapify(int i);
 
-	long long parent(long long i) { return (i-1)/2; }
+	int parent(int i) { return (i-1)/2; }
 
 	// to get index of left child of node at index i
-	long long left(long long i) { return (2*i + 1); }
+	int left(int i) { return (2*i + 1); }
 
 	// to get index of right child of node at index i
-	long long right(long long i) { return (2*i + 2); }
+	int right(int i) { return (2*i + 2); }
 
 	// to get the size of the current heap
-	long long getHeapSize() { return heap_size; };
+	int getHeapSize() { return heap_size; };
 
 	// to extract the root which is the minimum element
-	long long extractMin();
+	int extractMin();
 
 	// Decreases key value of key at index i to new_val
-	void decreaseKey(long long i, long long new_val);
+	void decreaseKey(int i, int new_val);
 
 	// Returns the minimum key (key at root) from min heap
-	long long getMin() { return harr[0]; }
+	int getMin() { return harr[0]; }
 
 	// Deletes a key stored at index i
-	void deleteKey(long long i);
+	void deleteKey(int i);
 
 	// Inserts a new key 'k'
-	void insertKey(long long k);
+	void insertKey(int k);
 };
 
 // Constructor: Builds a heap from a given array a[] of given size
-MinHeap::MinHeap(long long cap)
+MinHeap::MinHeap(int cap)
 {
 	heap_size = 0;
 	capacity = cap;
-	harr = new long long[cap];
+	harr = new int[cap];
 }
 
 // Inserts a new key 'k'
-void MinHeap::insertKey(long long k)
+void MinHeap::insertKey(int k)
 {
 	if (heap_size == capacity)
 	{
@@ -65,7 +65,7 @@ void MinHeap::insertKey(long long k)
 
 	// First insert the new key at the end
 	heap_size++;
-	long long i = heap_size - 1;
+	int i = heap_size - 1;
 	harr[i] = k;
 
 	// Fix the min heap property if it is violated
@@ -78,7 +78,7 @@ void MinHeap::insertKey(long long k)
 
 // Decreases value of key at index 'i' to new_val. It is assumed that
 // new_val is smaller than harr[i].
-void MinHeap::decreaseKey(long long i, long long new_val)
+void MinHeap::decreaseKey(int i, int new_val)
 {
 	harr[i] = new_val;
 	while (i != 0 && harr[parent(i)] > harr[i])
@@ -89,7 +89,7 @@ void MinHeap::decreaseKey(long long i, long long new_val)
 }
 
 // Method to remove minimum element (or root) from min heap
-long long MinHeap::extractMin()
+int MinHeap::extractMin()
 {
 	if (heap_size <= 0)
 		return LLONG_MAX;
@@ -111,7 +111,7 @@ long long MinHeap::extractMin()
 
 // This function deletes key at index i. It first reduced value to minus
 // infinite, then calls extractMin()
-void MinHeap::deleteKey(long long i)
+void MinHeap::deleteKey(int i)
 {
 	decreaseKey(i, LLONG_MAX);
 	extractMin();
@@ -119,11 +119,11 @@ void MinHeap::deleteKey(long long i)
 
 // A recursive method to heapify a subtree with the root at given index
 // This method assumes that the subtrees are already heapified
-void MinHeap::MinHeapify(long long i)
+void MinHeap::MinHeapify(int i)
 {
-	long long l = left(i);
-	long long r = right(i);
-	long long smallest = i;
+	int l = left(i);
+	int r = right(i);
+	int smallest = i;
 	if (l < heap_size && harr[l] < harr[i])
 		smallest = l;
 	if (r < heap_size && harr[r] < harr[smallest])
@@ -136,9 +136,9 @@ void MinHeap::MinHeapify(long long i)
 }
 
 // A utility function to swap two elements
-void swap(long long *x, long long *y)
+void swap(int *x, int *y)
 {
-	long long temp = *x;
+	int temp = *x;
 	*x = *y;
 	*y = temp;
 }
